COMMERCE-7204 Build Service

+++ modules/apps/commerce/commerce-product-service/src/main/java/com/liferay/commerce/product/service/persistence/impl/CPDefinitionLocalizationPersistenceImpl.java
@@ -22,6 +22,7 @@ import com.liferay.commerce.product.model.impl.CPDefinitionLocalizationModelImpl
 import com.liferay.commerce.product.service.persistence.CPDefinitionLocalizationPersistence;
 import com.liferay.commerce.product.service.persistence.CPDefinitionLocalizationUtil;
 import com.liferay.petra.string.StringBundler;
+import com.liferay.portal.kernel.change.tracking.CTColumnResolutionType;
 import com.liferay.portal.kernel.dao.orm.EntityCache;
 import com.liferay.portal.kernel.dao.orm.FinderCache;
 import com.liferay.portal.kernel.dao.orm.FinderPath;
@@ -37,6 +38,7 @@ import com.liferay.portal.kernel.sanitizer.SanitizerException;
 import com.liferay.portal.kernel.sanitizer.SanitizerUtil;
 import com.liferay.portal.kernel.security.auth.CompanyThreadLocal;
 import com.liferay.portal.kernel.security.auth.PrincipalThreadLocal;
+import com.liferay.portal.kernel.service.persistence.change.tracking.helper.CTPersistenceHelper;
 import com.liferay.portal.kernel.service.persistence.impl.BasePersistenceImpl;
 import com.liferay.portal.kernel.util.ContentTypes;
 import com.liferay.portal.kernel.util.GetterUtil;
@@ -51,6 +53,12 @@ import java.io.Serializable;
 import java.lang.reflect.Field;
 import java.lang.reflect.InvocationHandler;
 
+import java.util.ArrayList;
+import java.util.Collections;
+import java.util.EnumMap;
+import java.util.HashMap;
+import java.util.HashSet;
+import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
@@ -166,18 +174,21 @@ public class CPDefinitionLocalizationPersistenceImpl
 		OrderByComparator<CPDefinitionLocalization> orderByComparator,
 		boolean useFinderCache) {
 
+		boolean productionMode = ctPersistenceHelper.isProductionMode(
+			CPDefinitionLocalization.class);
+
 		FinderPath finderPath = null;
 		Object[] finderArgs = null;
 
 		if ((start == QueryUtil.ALL_POS) && (end == QueryUtil.ALL_POS) &&
 			(orderByComparator == null)) {
 
-			if (useFinderCache) {
+			if (useFinderCache && productionMode) {
 				finderPath = _finderPathWithoutPaginationFindByCPDefinitionId;
 				finderArgs = new Object[] {CPDefinitionId};
 			}
 		}
-		else if (useFinderCache) {
+		else if (useFinderCache && productionMode) {
 			finderPath = _finderPathWithPaginationFindByCPDefinitionId;
 			finderArgs = new Object[] {
 				CPDefinitionId, start, end, orderByComparator
@@ -186,7 +197,7 @@ public class CPDefinitionLocalizationPersistenceImpl
 
 		List<CPDefinitionLocalization> list = null;
 
-		if (useFinderCache) {
+		if (useFinderCache && productionMode) {
 			list = (List<CPDefinitionLocalization>)finderCache.getResult(
 				finderPath, finderArgs);
 
@@ -244,7 +255,7 @@ public class CPDefinitionLocalizationPersistenceImpl
 
 				cacheResult(list);
 
-				if (useFinderCache) {
+				if (useFinderCache && productionMode) {
 					finderCache.putResult(finderPath, finderArgs, list);
 				}
 			}
@@ -556,11 +567,21 @@ public class CPDefinitionLocalizationPersistenceImpl
 	 */
 	@Override
 	public int countByCPDefinitionId(long CPDefinitionId) {
-		FinderPath finderPath = _finderPathCountByCPDefinitionId;
+		boolean productionMode = ctPersistenceHelper.isProductionMode(
+			CPDefinitionLocalization.class);
 
-		Object[] finderArgs = new Object[] {CPDefinitionId};
+		FinderPath finderPath = null;
+		Object[] finderArgs = null;
 
-		Long count = (Long)finderCache.getResult(finderPath, finderArgs);
+		Long count = null;
+
+		if (productionMode) {
+			finderPath = _finderPathCountByCPDefinitionId;
+
+			finderArgs = new Object[] {CPDefinitionId};
+
+			count = (Long)finderCache.getResult(finderPath, finderArgs);
+		}
 
 		if (count == null) {
 			StringBundler sb = new StringBundler(2);
@@ -584,7 +605,9 @@ public class CPDefinitionLocalizationPersistenceImpl
 
 				count = (Long)query.uniqueResult();
 
-				finderCache.putResult(finderPath, finderArgs, count);
+				if (productionMode) {
+					finderCache.putResult(finderPath, finderArgs, count);
+				}
 			}
 			catch (Exception exception) {
 				throw processException(exception);
@@ -671,15 +694,18 @@ public class CPDefinitionLocalizationPersistenceImpl
 
 		languageId = Objects.toString(languageId, "");
 
+		boolean productionMode = ctPersistenceHelper.isProductionMode(
+			CPDefinitionLocalization.class);
+
 		Object[] finderArgs = null;
 
-		if (useFinderCache) {
+		if (useFinderCache && productionMode) {
 			finderArgs = new Object[] {CPDefinitionId, languageId};
 		}
 
 		Object result = null;
 
-		if (useFinderCache) {
+		if (useFinderCache && productionMode) {
 			result = finderCache.getResult(
 				_finderPathFetchByCPDefinitionId_LanguageId, finderArgs);
 		}
@@ -738,7 +764,7 @@ public class CPDefinitionLocalizationPersistenceImpl
 				List<CPDefinitionLocalization> list = query.list();
 
 				if (list.isEmpty()) {
-					if (useFinderCache) {
+					if (useFinderCache && productionMode) {
 						finderCache.putResult(
 							_finderPathFetchByCPDefinitionId_LanguageId,
 							finderArgs, list);
@@ -800,11 +826,21 @@ public class CPDefinitionLocalizationPersistenceImpl
 
 		languageId = Objects.toString(languageId, "");
 
-		FinderPath finderPath = _finderPathCountByCPDefinitionId_LanguageId;
+		boolean productionMode = ctPersistenceHelper.isProductionMode(
+			CPDefinitionLocalization.class);
 
-		Object[] finderArgs = new Object[] {CPDefinitionId, languageId};
+		FinderPath finderPath = null;
+		Object[] finderArgs = null;
 
-		Long count = (Long)finderCache.getResult(finderPath, finderArgs);
+		Long count = null;
+
+		if (productionMode) {
+			finderPath = _finderPathCountByCPDefinitionId_LanguageId;
+
+			finderArgs = new Object[] {CPDefinitionId, languageId};
+
+			count = (Long)finderCache.getResult(finderPath, finderArgs);
+		}
 
 		if (count == null) {
 			StringBundler sb = new StringBundler(3);
@@ -846,7 +882,9 @@ public class CPDefinitionLocalizationPersistenceImpl
 
 				count = (Long)query.uniqueResult();
 
-				finderCache.putResult(finderPath, finderArgs, count);
+				if (productionMode) {
+					finderCache.putResult(finderPath, finderArgs, count);
+				}
 			}
 			catch (Exception exception) {
 				throw processException(exception);
@@ -887,6 +925,10 @@ public class CPDefinitionLocalizationPersistenceImpl
 	 */
 	@Override
 	public void cacheResult(CPDefinitionLocalization cpDefinitionLocalization) {
+		if (cpDefinitionLocalization.getCtCollectionId() != 0) {
+			return;
+		}
+
 		entityCache.putResult(
 			CPDefinitionLocalizationImpl.class,
 			cpDefinitionLocalization.getPrimaryKey(), cpDefinitionLocalization);
@@ -921,6 +963,10 @@ public class CPDefinitionLocalizationPersistenceImpl
 
 		for (CPDefinitionLocalization cpDefinitionLocalization :
 				cpDefinitionLocalizations) {
+
+			if (cpDefinitionLocalization.getCtCollectionId() != 0) {
+				continue;
+			}
 
 			if (entityCache.getResult(
 					CPDefinitionLocalizationImpl.class,
@@ -1087,7 +1133,9 @@ public class CPDefinitionLocalizationPersistenceImpl
 						cpDefinitionLocalization.getPrimaryKeyObj());
 			}
 
-			if (cpDefinitionLocalization != null) {
+			if ((cpDefinitionLocalization != null) &&
+				ctPersistenceHelper.isRemove(cpDefinitionLocalization)) {
+
 				session.delete(cpDefinitionLocalization);
 			}
 		}
@@ -1166,7 +1214,13 @@ public class CPDefinitionLocalizationPersistenceImpl
 		try {
 			session = openSession();
 
-			if (isNew) {
+			if (ctPersistenceHelper.isInsert(cpDefinitionLocalization)) {
+				if (!isNew) {
+					session.evict(
+						CPDefinitionLocalizationImpl.class,
+						cpDefinitionLocalization.getPrimaryKeyObj());
+				}
+
 				session.save(cpDefinitionLocalization);
 			}
 			else {
@@ -1180,6 +1234,16 @@ public class CPDefinitionLocalizationPersistenceImpl
 		}
 		finally {
 			closeSession(session);
+		}
+
+		if (cpDefinitionLocalization.getCtCollectionId() != 0) {
+			if (isNew) {
+				cpDefinitionLocalization.setNew(false);
+			}
+
+			cpDefinitionLocalization.resetOriginalValues();
+
+			return cpDefinitionLocalization;
 		}
 
 		entityCache.putResult(
@@ -1241,6 +1305,44 @@ public class CPDefinitionLocalizationPersistenceImpl
 	/**
 	 * Returns the cp definition localization with the primary key or returns <code>null</code> if it could not be found.
 	 *
+	 * @param primaryKey the primary key of the cp definition localization
+	 * @return the cp definition localization, or <code>null</code> if a cp definition localization with the primary key could not be found
+	 */
+	@Override
+	public CPDefinitionLocalization fetchByPrimaryKey(Serializable primaryKey) {
+		if (ctPersistenceHelper.isProductionMode(
+				CPDefinitionLocalization.class)) {
+
+			return super.fetchByPrimaryKey(primaryKey);
+		}
+
+		CPDefinitionLocalization cpDefinitionLocalization = null;
+
+		Session session = null;
+
+		try {
+			session = openSession();
+
+			cpDefinitionLocalization = (CPDefinitionLocalization)session.get(
+				CPDefinitionLocalizationImpl.class, primaryKey);
+
+			if (cpDefinitionLocalization != null) {
+				cacheResult(cpDefinitionLocalization);
+			}
+		}
+		catch (Exception exception) {
+			throw processException(exception);
+		}
+		finally {
+			closeSession(session);
+		}
+
+		return cpDefinitionLocalization;
+	}
+
+	/**
+	 * Returns the cp definition localization with the primary key or returns <code>null</code> if it could not be found.
+	 *
 	 * @param cpDefinitionLocalizationId the primary key of the cp definition localization
 	 * @return the cp definition localization, or <code>null</code> if a cp definition localization with the primary key could not be found
 	 */
@@ -1249,6 +1351,104 @@ public class CPDefinitionLocalizationPersistenceImpl
 		long cpDefinitionLocalizationId) {
 
 		return fetchByPrimaryKey((Serializable)cpDefinitionLocalizationId);
+	}
+
+	@Override
+	public Map<Serializable, CPDefinitionLocalization> fetchByPrimaryKeys(
+		Set<Serializable> primaryKeys) {
+
+		if (ctPersistenceHelper.isProductionMode(
+				CPDefinitionLocalization.class)) {
+
+			return super.fetchByPrimaryKeys(primaryKeys);
+		}
+
+		if (primaryKeys.isEmpty()) {
+			return Collections.emptyMap();
+		}
+
+		Map<Serializable, CPDefinitionLocalization> map =
+			new HashMap<Serializable, CPDefinitionLocalization>();
+
+		if (primaryKeys.size() == 1) {
+			Iterator<Serializable> iterator = primaryKeys.iterator();
+
+			Serializable primaryKey = iterator.next();
+
+			CPDefinitionLocalization cpDefinitionLocalization =
+				fetchByPrimaryKey(primaryKey);
+
+			if (cpDefinitionLocalization != null) {
+				map.put(primaryKey, cpDefinitionLocalization);
+			}
+
+			return map;
+		}
+
+		if ((databaseInMaxParameters > 0) &&
+			(primaryKeys.size() > databaseInMaxParameters)) {
+
+			Iterator<Serializable> iterator = primaryKeys.iterator();
+
+			while (iterator.hasNext()) {
+				Set<Serializable> page = new HashSet<>();
+
+				for (int i = 0;
+					 (i < databaseInMaxParameters) && iterator.hasNext(); i++) {
+
+					page.add(iterator.next());
+				}
+
+				map.putAll(fetchByPrimaryKeys(page));
+			}
+
+			return map;
+		}
+
+		StringBundler sb = new StringBundler((primaryKeys.size() * 2) + 1);
+
+		sb.append(getSelectSQL());
+		sb.append(" WHERE ");
+		sb.append(getPKDBName());
+		sb.append(" IN (");
+
+		for (Serializable primaryKey : primaryKeys) {
+			sb.append((long)primaryKey);
+
+			sb.append(",");
+		}
+
+		sb.setIndex(sb.index() - 1);
+
+		sb.append(")");
+
+		String sql = sb.toString();
+
+		Session session = null;
+
+		try {
+			session = openSession();
+
+			Query query = session.createQuery(sql);
+
+			for (CPDefinitionLocalization cpDefinitionLocalization :
+					(List<CPDefinitionLocalization>)query.list()) {
+
+				map.put(
+					cpDefinitionLocalization.getPrimaryKeyObj(),
+					cpDefinitionLocalization);
+
+				cacheResult(cpDefinitionLocalization);
+			}
+		}
+		catch (Exception exception) {
+			throw processException(exception);
+		}
+		finally {
+			closeSession(session);
+		}
+
+		return map;
 	}
 
 	/**
@@ -1316,25 +1516,28 @@ public class CPDefinitionLocalizationPersistenceImpl
 		OrderByComparator<CPDefinitionLocalization> orderByComparator,
 		boolean useFinderCache) {
 
+		boolean productionMode = ctPersistenceHelper.isProductionMode(
+			CPDefinitionLocalization.class);
+
 		FinderPath finderPath = null;
 		Object[] finderArgs = null;
 
 		if ((start == QueryUtil.ALL_POS) && (end == QueryUtil.ALL_POS) &&
 			(orderByComparator == null)) {
 
-			if (useFinderCache) {
+			if (useFinderCache && productionMode) {
 				finderPath = _finderPathWithoutPaginationFindAll;
 				finderArgs = FINDER_ARGS_EMPTY;
 			}
 		}
-		else if (useFinderCache) {
+		else if (useFinderCache && productionMode) {
 			finderPath = _finderPathWithPaginationFindAll;
 			finderArgs = new Object[] {start, end, orderByComparator};
 		}
 
 		List<CPDefinitionLocalization> list = null;
 
-		if (useFinderCache) {
+		if (useFinderCache && productionMode) {
 			list = (List<CPDefinitionLocalization>)finderCache.getResult(
 				finderPath, finderArgs);
 		}
@@ -1373,7 +1576,7 @@ public class CPDefinitionLocalizationPersistenceImpl
 
 				cacheResult(list);
 
-				if (useFinderCache) {
+				if (useFinderCache && productionMode) {
 					finderCache.putResult(finderPath, finderArgs, list);
 				}
 			}
@@ -1406,8 +1609,15 @@ public class CPDefinitionLocalizationPersistenceImpl
 	 */
 	@Override
 	public int countAll() {
-		Long count = (Long)finderCache.getResult(
-			_finderPathCountAll, FINDER_ARGS_EMPTY);
+		boolean productionMode = ctPersistenceHelper.isProductionMode(
+			CPDefinitionLocalization.class);
+
+		Long count = null;
+
+		if (productionMode) {
+			count = (Long)finderCache.getResult(
+				_finderPathCountAll, FINDER_ARGS_EMPTY);
+		}
 
 		if (count == null) {
 			Session session = null;
@@ -1420,8 +1630,10 @@ public class CPDefinitionLocalizationPersistenceImpl
 
 				count = (Long)query.uniqueResult();
 
-				finderCache.putResult(
-					_finderPathCountAll, FINDER_ARGS_EMPTY, count);
+				if (productionMode) {
+					finderCache.putResult(
+						_finderPathCountAll, FINDER_ARGS_EMPTY, count);
+				}
 			}
 			catch (Exception exception) {
 				throw processException(exception);
@@ -1450,8 +1662,67 @@ public class CPDefinitionLocalizationPersistenceImpl
 	}
 
 	@Override
-	protected Map<String, Integer> getTableColumnsMap() {
+	public Set<String> getCTColumnNames(
+		CTColumnResolutionType ctColumnResolutionType) {
+
+		return _ctColumnNamesMap.getOrDefault(
+			ctColumnResolutionType, Collections.emptySet());
+	}
+
+	@Override
+	public List<String> getMappingTableNames() {
+		return _mappingTableNames;
+	}
+
+	@Override
+	public Map<String, Integer> getTableColumnsMap() {
 		return CPDefinitionLocalizationModelImpl.TABLE_COLUMNS_MAP;
+	}
+
+	@Override
+	public String getTableName() {
+		return "CPDefinitionLocalization";
+	}
+
+	@Override
+	public List<String[]> getUniqueIndexColumnNames() {
+		return _uniqueIndexColumnNames;
+	}
+
+	private static final Map<CTColumnResolutionType, Set<String>>
+		_ctColumnNamesMap = new EnumMap<CTColumnResolutionType, Set<String>>(
+			CTColumnResolutionType.class);
+	private static final List<String> _mappingTableNames =
+		new ArrayList<String>();
+	private static final List<String[]> _uniqueIndexColumnNames =
+		new ArrayList<String[]>();
+
+	static {
+		Set<String> ctControlColumnNames = new HashSet<String>();
+		Set<String> ctStrictColumnNames = new HashSet<String>();
+
+		ctControlColumnNames.add("mvccVersion");
+		ctControlColumnNames.add("ctCollectionId");
+		ctStrictColumnNames.add("companyId");
+		ctStrictColumnNames.add("CPDefinitionId");
+		ctStrictColumnNames.add("languageId");
+		ctStrictColumnNames.add("name");
+		ctStrictColumnNames.add("shortDescription");
+		ctStrictColumnNames.add("description");
+		ctStrictColumnNames.add("metaTitle");
+		ctStrictColumnNames.add("metaDescription");
+		ctStrictColumnNames.add("metaKeywords");
+
+		_ctColumnNamesMap.put(
+			CTColumnResolutionType.CONTROL, ctControlColumnNames);
+		_ctColumnNamesMap.put(
+			CTColumnResolutionType.PK,
+			Collections.singleton("cpDefinitionLocalizationId"));
+		_ctColumnNamesMap.put(
+			CTColumnResolutionType.STRICT, ctStrictColumnNames);
+
+		_uniqueIndexColumnNames.add(
+			new String[] {"CPDefinitionId", "languageId"});
 	}
 
 	/**
@@ -1527,6 +1798,9 @@ public class CPDefinitionLocalizationPersistenceImpl
 			throw new RuntimeException(reflectiveOperationException);
 		}
 	}
+
+	@ServiceReference(type = CTPersistenceHelper.class)
+	protected CTPersistenceHelper ctPersistenceHelper;
 
 	@ServiceReference(type = EntityCache.class)
 	protected EntityCache entityCache;
